+++ core/java/androidx/test/core/view/ViewCapture.kt
@@ -32,17 +32,24 @@ import android.view.View
 import android.view.ViewTreeObserver
 import android.view.WindowManager
 import androidx.annotation.RequiresApi
-import androidx.concurrent.futures.ResolvableFuture
+import androidx.concurrent.futures.SuspendToFutureAdapter
 import androidx.test.annotation.ExperimentalTestApi
 import androidx.test.core.internal.os.HandlerExecutor
+import androidx.test.internal.platform.ServiceLoaderWrapper
+import androidx.test.internal.platform.os.ControlledLooper
 import androidx.test.internal.platform.reflect.ReflectiveField
 import androidx.test.internal.platform.reflect.ReflectiveMethod
 import androidx.test.platform.graphics.HardwareRendererCompat
 import com.google.common.util.concurrent.ListenableFuture
 import java.util.function.Consumer
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.android.asCoroutineDispatcher
+import kotlinx.coroutines.launch
 
 /**
- * Asynchronously captures an image of the underlying view into a [Bitmap].
+ * Suspend function for capturing an image of the underlying view into a [Bitmap].
  *
  * For devices below [Build.VERSION_CODES#O] (or if the view's window cannot be determined), the
  * image is obtained using [View#draw]. Otherwise, [PixelCopy] is used.
@@ -50,7 +57,7 @@ import java.util.function.Consumer
  * This method will also enable [HardwareRendererCompat#setDrawingEnabled(boolean)] if required.
  *
  * This API is primarily intended for use in lower layer libraries or frameworks. For test authors,
- * its recommended to use espresso or compose's captureToImage.
+ * it's recommended to use Espresso or Compose's captureToImage.
  *
  * If a rect is supplied, this will further crop locally from the bounds of the given view. For
  * example, if the given view is at (10, 10 - 30, 30) and the rect is (5, 5 - 10, 10), the final
@@ -60,43 +67,52 @@ import java.util.function.Consumer
  * This API is currently experimental and subject to change or removal.
  */
 @ExperimentalTestApi
-fun View.captureToBitmap(rect: Rect? = null): ListenableFuture<Bitmap> {
-  val bitmapFuture: ResolvableFuture<Bitmap> = ResolvableFuture.create()
-  val mainExecutor = HandlerExecutor(Handler(Looper.getMainLooper()))
+suspend fun View.captureToBitmap(rect: Rect? = null): Bitmap {
+  val mainHandlerDispatcher = Handler(Looper.getMainLooper()).asCoroutineDispatcher()
+  var bitmap: Bitmap? = null
 
-  // disable drawing again if necessary once work is complete
-  if (!HardwareRendererCompat.isDrawingEnabled()) {
-    HardwareRendererCompat.setDrawingEnabled(true)
-    bitmapFuture.addListener({ HardwareRendererCompat.setDrawingEnabled(false) }, mainExecutor)
-  }
-
-  mainExecutor.execute {
-    if (Build.FINGERPRINT.contains("robolectric")) {
-      generateBitmap(bitmapFuture, rect)
-    } else {
-      val forceRedrawFuture = forceRedraw()
-      forceRedrawFuture.addListener({ generateBitmap(bitmapFuture, rect) }, mainExecutor)
+  val job =
+    CoroutineScope(mainHandlerDispatcher).launch {
+      val hardwareDrawingEnabled = HardwareRendererCompat.isDrawingEnabled()
+      HardwareRendererCompat.setDrawingEnabled(true)
+      forceRedraw()
+      bitmap = generateBitmap(rect)
+      HardwareRendererCompat.setDrawingEnabled(hardwareDrawingEnabled)
     }
-  }
 
-  return bitmapFuture
+  getControlledLooper().drainMainThreadUntilIdle()
+  job.join()
+
+  return bitmap!!
+}
+
+private fun getControlledLooper(): ControlledLooper {
+  return ServiceLoaderWrapper.loadSingleService(ControlledLooper::class.java) {
+    ControlledLooper.NO_OP_CONTROLLED_LOOPER
+  }
+}
+
+/** A ListenableFuture variant of captureToBitmap intended for use from Java. */
+@ExperimentalTestApi
+fun View.captureToBitmapAsync(rect: Rect? = null): ListenableFuture<Bitmap> {
+  return SuspendToFutureAdapter.launchFuture(Dispatchers.Default + Job()) { captureToBitmap(rect) }
 }
 
 /**
  * Trigger a redraw of the given view.
  *
  * Should only be called on UI thread.
- *
- * @return a [ListenableFuture] that will be complete once ui drawing is complete
  */
 // TODO(b/316921934): uncomment once @ExperimentalTestApi is removed
 // @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 @ExperimentalTestApi
-fun View.forceRedraw(): ListenableFuture<Void> {
-  val future: ResolvableFuture<Void> = ResolvableFuture.create()
-
+suspend fun View.forceRedraw() {
+  if (!getControlledLooper().isDrawCallbacksSupported) {
+    Log.i("ViewCapture", "Skipping isDrawCallbacksSupported as it is not supported")
+  }
+  val job = Job()
   if (Build.VERSION.SDK_INT >= 29 && isHardwareAccelerated) {
-    viewTreeObserver.registerFrameCommitCallback() { future.set(null) }
+    viewTreeObserver.registerFrameCommitCallback() { job.complete() }
   } else {
     viewTreeObserver.addOnDrawListener(
       object : ViewTreeObserver.OnDrawListener {
@@ -105,7 +121,7 @@ fun View.forceRedraw(): ListenableFuture<Void> {
         override fun onDraw() {
           if (!handled) {
             handled = true
-            future.set(null)
+            job.complete()
             // cannot remove on draw listener inside of onDraw
             Handler(Looper.getMainLooper()).post { viewTreeObserver.removeOnDrawListener(this) }
           }
@@ -114,46 +130,47 @@ fun View.forceRedraw(): ListenableFuture<Void> {
     )
   }
   invalidate()
-  return future
+  job.join()
 }
 
-private fun View.generateBitmap(bitmapFuture: ResolvableFuture<Bitmap>, rect: Rect? = null) {
-  if (bitmapFuture.isCancelled) {
-    return
-  }
+private suspend fun View.generateBitmap(rect: Rect? = null): Bitmap {
   val rectWidth = rect?.width() ?: width
   val rectHeight = rect?.height() ?: height
   val destBitmap = Bitmap.createBitmap(rectWidth, rectHeight, Bitmap.Config.ARGB_8888)
-  when {
-    Build.VERSION.SDK_INT < 26 -> generateBitmapFromDraw(destBitmap, bitmapFuture, rect)
-    Build.VERSION.SDK_INT >= 34 -> generateBitmapFromPixelCopy(destBitmap, bitmapFuture, rect)
-    this is SurfaceView -> generateBitmapFromSurfaceViewPixelCopy(destBitmap, bitmapFuture, rect)
-    else -> generateBitmapFromPixelCopy(this.getSurface(), destBitmap, bitmapFuture, rect)
+
+  return when {
+    Build.VERSION.SDK_INT < 26 -> generateBitmapFromDraw(destBitmap, rect)
+    Build.VERSION.SDK_INT >= 34 -> generateBitmapFromPixelCopy(destBitmap, rect)
+    this is SurfaceView -> generateBitmapFromSurfaceViewPixelCopy(destBitmap, rect)
+    else -> generateBitmapFromPixelCopy(this.getSurface(), destBitmap, rect)
   }
 }
 
 @RequiresApi(Build.VERSION_CODES.O)
-private fun SurfaceView.generateBitmapFromSurfaceViewPixelCopy(
+private suspend fun SurfaceView.generateBitmapFromSurfaceViewPixelCopy(
   destBitmap: Bitmap,
-  bitmapFuture: ResolvableFuture<Bitmap>,
   rect: Rect?,
-) {
+): Bitmap {
+  val job = Job()
+  var exception: Exception? = null
+  var bitmap: Bitmap? = null
   val onCopyFinished =
     PixelCopy.OnPixelCopyFinishedListener { result ->
       if (result == PixelCopy.SUCCESS) {
-        bitmapFuture.set(destBitmap)
+        bitmap = destBitmap
       } else {
-        bitmapFuture.setException(RuntimeException(String.format("PixelCopy failed: %d", result)))
+        exception = RuntimeException(String.format("PixelCopy failed: %d", result))
       }
+      job.complete()
     }
   PixelCopy.request(this, rect, destBitmap, onCopyFinished, handler)
+
+  job.join()
+  exception?.let { throw it }
+  return bitmap!!
 }
 
-internal fun View.generateBitmapFromDraw(
-  destBitmap: Bitmap,
-  bitmapFuture: ResolvableFuture<Bitmap>,
-  rect: Rect?,
-) {
+internal suspend fun View.generateBitmapFromDraw(destBitmap: Bitmap, rect: Rect?): Bitmap {
   destBitmap.density = resources.displayMetrics.densityDpi
   computeScroll()
   val canvas = Canvas(destBitmap)
@@ -163,7 +180,7 @@ internal fun View.generateBitmapFromDraw(
   }
 
   draw(canvas)
-  bitmapFuture.set(destBitmap)
+  return destBitmap
 }
 
 /**
@@ -173,19 +190,23 @@ internal fun View.generateBitmapFromDraw(
  * [PixelCopy.ofWindow(View)], and will be called when running on Android API levels O to T.
  */
 @RequiresApi(Build.VERSION_CODES.O)
-private fun View.generateBitmapFromPixelCopy(
+private suspend fun View.generateBitmapFromPixelCopy(
   surface: Surface,
   destBitmap: Bitmap,
-  bitmapFuture: ResolvableFuture<Bitmap>,
   rect: Rect?,
-) {
+): Bitmap {
+  val job = Job()
+  var exception: Exception? = null
+  var bitmap: Bitmap? = null
+
   val onCopyFinished =
     PixelCopy.OnPixelCopyFinishedListener { result ->
       if (result == PixelCopy.SUCCESS) {
-        bitmapFuture.set(destBitmap)
+        bitmap = destBitmap
       } else {
-        bitmapFuture.setException(RuntimeException("PixelCopy failed: $result"))
+        exception = RuntimeException("PixelCopy failed: $result")
       }
+      job.complete()
     }
 
   var bounds = getBoundsInSurface()
@@ -195,10 +216,13 @@ private fun View.generateBitmapFromPixelCopy(
         bounds.left + rect.left,
         bounds.top + rect.top,
         bounds.left + rect.right,
-        bounds.top + rect.bottom
+        bounds.top + rect.bottom,
       )
   }
   PixelCopy.request(surface, bounds, destBitmap, onCopyFinished, Handler(Looper.getMainLooper()))
+  job.join()
+  exception?.let { throw it }
+  return bitmap!!
 }
 
 /** Returns the Rect indicating the View's coordinates within its containing window. */
@@ -256,32 +280,35 @@ private fun View.reflectivelyGetLocationInSurface(locationInSurface: IntArray) {
     // ART restrictions introduced in API 29 disallow reflective access to mWindowAttributes
     Log.w(
       "ViewCapture",
-      "Could not calculate offset of view in surface on API 28, resulting image may have incorrect positioning"
+      "Could not calculate offset of view in surface on API 28, resulting image may have incorrect positioning",
     )
   }
 }
 
-/** Generates a bitmap given the current view using android U's [PixelCopy.ofWindow(View)]. */
-@RequiresApi(Build.VERSION_CODES.UPSIDE_DOWN_CAKE)
-private fun View.generateBitmapFromPixelCopy(
+private suspend fun View.generateBitmapFromPixelCopy(
   destBitmap: Bitmap,
-  bitmapFuture: ResolvableFuture<Bitmap>,
-  rect: Rect?,
-) {
+  rect: Rect? = null,
+): Bitmap {
+  val job = Job()
+  var exception: Exception? = null
   val request =
     PixelCopy.Request.Builder.ofWindow(this)
       .setSourceRect(rect ?: getBoundsInWindow())
       .setDestinationBitmap(destBitmap)
       .build()
   val mainExecutor = HandlerExecutor(Handler(Looper.getMainLooper()))
-
+  var ret: Bitmap? = null
   val onCopyFinished =
     Consumer<PixelCopy.Result> { result ->
       if (result.status == PixelCopy.SUCCESS) {
-        bitmapFuture.set(result.bitmap)
+        ret = result.bitmap
       } else {
-        bitmapFuture.setException(RuntimeException("PixelCopy failed: $(result.status)"))
+        exception = RuntimeException("PixelCopy failed: $(result.status)")
       }
+      job.complete()
     }
   PixelCopy.request(request, mainExecutor, onCopyFinished)
+  job.join()
+  exception?.let { throw it }
+  return ret!!
 }