+++ core/java/androidx/test/core/view/WindowCapture.kt
@@ -24,14 +24,18 @@ import android.os.Handler
 import android.os.Looper
 import android.view.PixelCopy
 import android.view.Window
-import androidx.concurrent.futures.ResolvableFuture
+import androidx.concurrent.futures.SuspendToFutureAdapter
 import androidx.test.annotation.ExperimentalTestApi
-import androidx.test.core.internal.os.HandlerExecutor
 import androidx.test.platform.graphics.HardwareRendererCompat
 import com.google.common.util.concurrent.ListenableFuture
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.android.asCoroutineDispatcher
+import kotlinx.coroutines.launch
 
 /**
- * Asynchronously captures an image of the underlying window into a [Bitmap].
+ * Suspend function that captures an image of the underlying window into a [Bitmap].
  *
  * For devices below [Build.VERSION_CODES#O] the image is obtained using [View#draw] on the windows
  * decorView. Otherwise, [PixelCopy] is used.
@@ -44,61 +48,74 @@ import com.google.common.util.concurrent.ListenableFuture
  * This API is currently experimental and subject to change or removal.
  */
 @ExperimentalTestApi
-fun Window.captureRegionToBitmap(boundsInWindow: Rect? = null): ListenableFuture<Bitmap> {
-  val bitmapFuture: ResolvableFuture<Bitmap> = ResolvableFuture.create()
-  val mainExecutor = HandlerExecutor(Handler(Looper.getMainLooper()))
+suspend fun Window.captureRegionToBitmap(boundsInWindow: Rect? = null): Bitmap {
+  val mainHandlerDispatcher = Handler(Looper.getMainLooper()).asCoroutineDispatcher()
+  var bitmap: Bitmap? = null
 
-  // disable drawing again if necessary once work is complete
-  if (!HardwareRendererCompat.isDrawingEnabled()) {
-    HardwareRendererCompat.setDrawingEnabled(true)
-    bitmapFuture.addListener({ HardwareRendererCompat.setDrawingEnabled(false) }, mainExecutor)
-  }
+  val job =
+    CoroutineScope(mainHandlerDispatcher).launch {
+      val hardwareDrawingEnabled = HardwareRendererCompat.isDrawingEnabled()
+      HardwareRendererCompat.setDrawingEnabled(true)
+      if (getControlledLooper().isDrawCallbacksSupported) {
+        decorView.forceRedraw()
+      }
+      bitmap = generateBitmap(boundsInWindow)
+      HardwareRendererCompat.setDrawingEnabled(hardwareDrawingEnabled)
+    }
 
-  mainExecutor.execute {
-    val forceRedrawFuture = decorView.forceRedraw()
-    forceRedrawFuture.addListener({ generateBitmap(boundsInWindow, bitmapFuture) }, mainExecutor)
-  }
+  job.join()
 
-  return bitmapFuture
+  return bitmap!!
 }
 
-internal fun Window.generateBitmap(
-  boundsInWindow: Rect? = null,
-  bitmapFuture: ResolvableFuture<Bitmap>
-) {
+/** A ListenableFuture variant of captureRegionToBitmap intended for use from Java. */
+@ExperimentalTestApi
+fun Window.captureRegionToBitmapAsync(boundsInWindow: Rect? = null): ListenableFuture<Bitmap> {
+  return SuspendToFutureAdapter.launchFuture(Dispatchers.Default + Job()) {
+    captureRegionToBitmap(boundsInWindow)
+  }
+}
+
+internal suspend fun Window.generateBitmap(boundsInWindow: Rect? = null): Bitmap {
   val destBitmap =
     Bitmap.createBitmap(
       boundsInWindow?.width() ?: decorView.width,
       boundsInWindow?.height() ?: decorView.height,
-      Bitmap.Config.ARGB_8888
+      Bitmap.Config.ARGB_8888,
     )
   when {
     Build.VERSION.SDK_INT < 26 ->
       // TODO: handle boundsInWindow
-      decorView.generateBitmapFromDraw(destBitmap, bitmapFuture, null)
-    else -> generateBitmapFromPixelCopy(boundsInWindow, destBitmap, bitmapFuture)
+      decorView.generateBitmapFromDraw(destBitmap, boundsInWindow)
+    else -> generateBitmapFromPixelCopy(boundsInWindow, destBitmap)
   }
+
+  return destBitmap
 }
 
 @SuppressWarnings("NewApi")
-internal fun Window.generateBitmapFromPixelCopy(
+internal suspend fun Window.generateBitmapFromPixelCopy(
   boundsInWindow: Rect? = null,
   destBitmap: Bitmap,
-  bitmapFuture: ResolvableFuture<Bitmap>
 ) {
+  val job = Job()
+  var exception: Exception? = null
   val onCopyFinished =
     PixelCopy.OnPixelCopyFinishedListener { result ->
-      if (result == PixelCopy.SUCCESS) {
-        bitmapFuture.set(destBitmap)
-      } else {
-        bitmapFuture.setException(RuntimeException(String.format("PixelCopy failed: %d", result)))
+      if (result != PixelCopy.SUCCESS) {
+        exception = RuntimeException(String.format("PixelCopy failed: %d", result))
       }
+      job.complete()
     }
+
   PixelCopy.request(
     this,
     boundsInWindow,
     destBitmap,
     onCopyFinished,
-    Handler(Looper.getMainLooper())
+    Handler(Looper.getMainLooper()),
   )
+
+  job.join()
+  exception?.let { throw it }
 }